+++ src/Service/Sqs/src/Exception/BatchEntryIdsNotDistinctException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace AsyncAws\Sqs\Exception;
+
+use AsyncAws\Core\Exception\Http\ClientException;
+
+/**
+ * Two or more batch entries in the request have the same `Id`.
+ */
+final class BatchEntryIdsNotDistinctException extends ClientException
+{
+}