+++ src/Service/Sqs/CHANGELOG.md
@@ -6,7 +6,8 @@
 
 - AWS api-change: Added `us-iso-west-1` region
 - AWS api-change: Use specific configuration for `us` regions
-- AWS api-change: Amazon SQS adds a new queue attribute, SqsManagedSseEnabled, which enables server-side queue encryption using SQS owned encryption keys.
+- AWS api-change: Amazon SQS adds a new queue attribute, `SqsManagedSseEnabled`, which enables server-side queue encryption using SQS owned encryption keys.
+- Added operations `ChangeMessageVisibilityBatch`, `DeleteMessageBatch`, `SendMessageBatch`
 
 ## 1.6.0